pyproject.toml
[tool.poetry]
name = "devman-file-streamer"
version = "0.1.0"
description = "module 3 of devman's async python course"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.5"
aiofiles = "^0.4.0"
python-dotenv = "^0.10.1"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
pytest = "^4.4"
flake8 = "^3.7"
tox = "^3.9"
rope = "^0.14.0"
pytest-asyncio = "^0.10.0"
pytest-cov = "^2.6"
pytest-aiohttp = "^0.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
